requirements.txt
nbformat>=5.0.0
pyyaml>=6.0
openai>=1.0.0  # For OpenAI GPT integration
anthropic>=0.8.0  # For Claude integration
jupyter>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0